pyproject.toml
[project]
name = "ugc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Sergey Syrnikov", email = "sergeysyrnikov91@gmail.com" }
]
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.15",
    "prometheus-client>=0.24.1",
    "psycopg2>=2.9.11",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.9",
    "djangorestframework-stubs>=3.16.8",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.12.0",
]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = true
exclude = "migrations/"

[tool.django-stubs]
django_settings_module = "ugc.settings"
